Train on bid/ask does wont result in the same state
Approximation would be required in order to relate to a previously
visited state (see #4) and determine future reward. However, then lots of information gets lost
which could be of use for the NN.
One possible approach would be to focus on immediate reward and train on
optimal action for every execution step.